requirements.txt
Pillow==11.3.0
reportlab==4.4.5
tqdm==4.67.1